refactor(server): move model params into cli/server/types
The keep-alive cache keys on ModelParam / SpecParam, but they lived in
cli/internal/types (ModelParam) and cli/server/service (SpecParam),
straddling the server boundary. Move both into a leaf cli/server/types
package and fold the five flat SpecParam fields on ModelParam into a
single nested Spec field, so the cache key mirrors the request shape.

No behaviour change: reflect.DeepEqual over the nested struct compares
identically for cache-hit detection.
