+++ cli/server/handler/BUILD.bazel
@@ -20,8 +20,8 @@ go_library(
         "//cli/internal/config",
         "//cli/internal/store",
         "//cli/internal/thinkfsm",
-        "//cli/internal/types",
         "//cli/server/service",
+        "//cli/server/types",
         "//cli/server/utils",
         "@com_github_bytedance_sonic//:sonic",
         "@com_github_gin_gonic_gin//:gin",
