+++ cli/server/service/BUILD.bazel
@@ -13,8 +13,8 @@ go_library(
         "//bindings/go:geniex_sdk_go",
         "//cli/internal/config",
         "//cli/internal/render",
-        "//cli/internal/types",
         "//cli/server/middleware",
+        "//cli/server/types",
     ],
 )
