+++ cli/server/service/keepalive_test.go
@@ -9,6 +9,7 @@ import (
 
 	geniex_sdk "github.com/qualcomm/GenieX/bindings/go"
 	"github.com/qualcomm/GenieX/cli/server/middleware"
+	"github.com/qualcomm/GenieX/cli/server/types"
 )
 
 // ResolveModelParam receives already-resolved knobs (the handler prefills unset
@@ -18,7 +19,7 @@ import (
 // TestResolveModelParam_PassesLlamaCppValuesThrough verifies that nctx / ngl are
 // forwarded verbatim for llama_cpp and the compute alias resolves to a device.
 func TestResolveModelParam_PassesLlamaCppValuesThrough(t *testing.T) {
-	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 2048, 10, "gpu", "", SpecParam{})
+	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 2048, 10, "gpu", "", types.SpecParam{})
 	if err != nil {
 		t.Fatalf("ResolveModelParam: %v", err)
 	}
@@ -33,7 +34,7 @@ func TestResolveModelParam_PassesLlamaCppValuesThrough(t *testing.T) {
 // TestResolveModelParam_NpuAliasResolvesDevice verifies the npu alias pins HTP0
 // and passes ngl through (-1 = all layers).
 func TestResolveModelParam_NpuAliasResolvesDevice(t *testing.T) {
-	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 4096, -1, "npu", "", SpecParam{})
+	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 4096, -1, "npu", "", types.SpecParam{})
 	if err != nil {
 		t.Fatalf("ResolveModelParam: %v", err)
 	}
@@ -48,7 +49,7 @@ func TestResolveModelParam_NpuAliasResolvesDevice(t *testing.T) {
 // TestResolveModelParam_CpuAliasZeroesGpuLayers verifies ngl 0 (pure CPU) is a
 // valid value that survives resolution.
 func TestResolveModelParam_CpuAliasZeroesGpuLayers(t *testing.T) {
-	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 4096, 0, "cpu", "", SpecParam{})
+	got, err := ResolveModelParam(geniex_sdk.RuntimeLlamaCpp, "some-model", 4096, 0, "cpu", "", types.SpecParam{})
 	if err != nil {
 		t.Fatalf("ResolveModelParam: %v", err)
 	}
@@ -61,7 +62,7 @@ func TestResolveModelParam_CpuAliasZeroesGpuLayers(t *testing.T) {
 // runtimes NCtx is zeroed so the plugin's param-guard is not tripped, even when
 // the caller passes a non-zero value.
 func TestResolveModelParam_NonLlamaCppZeroesNCtx(t *testing.T) {
-	got, err := ResolveModelParam(geniex_sdk.RuntimeQairt, "some-model", 8192, 42, "", "", SpecParam{})
+	got, err := ResolveModelParam(geniex_sdk.RuntimeQairt, "some-model", 8192, 42, "", "", types.SpecParam{})
 	if err != nil {
 		t.Fatalf("ResolveModelParam: %v", err)
 	}
