+++ cli/server/types/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "types",
     srcs = ["model.go"],
-    importpath = "github.com/qualcomm/GenieX/cli/internal/types",
+    importpath = "github.com/qualcomm/GenieX/cli/server/types",
     visibility = ["//cli:__subpackages__"],
 )
